Introduce source file: invalidop.c
Logic for handling soft errors and displaying the invalid operation screen.
NOTE: The approach used in this file requires each source file that uses this function to implement a handler.
There may be multiple files, each with their own handler. This is not a problem as long as each handler function has a unique name - two or more functions with the same name will cause a compile or linking error.